Modularize popups into dedicated package
