+++ ui/popups/__init__.py
@@ -1,0 +1,11 @@
+from .constants import METRIC_FIELD_ORDER
+from .add_metric_popup import AddMetricPopup
+from .edit_metric_popup import EditMetricPopup
+from .pre_session_metric_popup import PreSessionMetricPopup
+
+__all__ = [
+    "AddMetricPopup",
+    "EditMetricPopup",
+    "PreSessionMetricPopup",
+    "METRIC_FIELD_ORDER",
+]
