+++ ui/popups/add_metric_popup.py
@@ -1,0 +1,321 @@
+from __future__ import annotations
+
+from kivymd.app import MDApp
+from kivy.metrics import dp
+from kivy.uix.spinner import Spinner
+from kivy.uix.scrollview import ScrollView
+from kivymd.uix.dialog import MDDialog
+from kivymd.uix.boxlayout import MDBoxLayout
+from kivymd.uix.textfield import MDTextField
+from kivymd.uix.selectioncontrol import MDCheckbox
+from kivymd.uix.button import MDRaisedButton
+from kivymd.uix.list import MDList, OneLineListItem
+from kivymd.uix.label import MDLabel
+from kivymd.uix.slider import MDSlider
+
+import string
+import re
+import sqlite3
+
+import core
+from core import DEFAULT_DB_PATH
+from .constants import METRIC_FIELD_ORDER
+
+class AddMetricPopup(MDDialog):
+    """Popup dialog for selecting or creating metrics."""
+
+    def __init__(
+        self,
+        screen: "EditExerciseScreen",
+        popup_mode: str = "select",
+        mode: str = "library",
+        **kwargs,
+    ):
+        self.screen = screen
+        self.mode = mode
+        self.popup_mode = popup_mode
+        if self.mode == "session":
+            content = MDBoxLayout()
+            close_btn = MDRaisedButton(
+                text="Close", on_release=lambda *a: self.dismiss()
+            )
+            super().__init__(
+                title="Metrics Locked",
+                type="custom",
+                content_cls=content,
+                buttons=[close_btn],
+                **kwargs,
+            )
+            return
+
+        if popup_mode == "select":
+            content, buttons, title = self._build_select_widgets()
+        elif popup_mode == "new":
+            content, buttons, title = self._build_new_metric_widgets()
+        else:
+            content, buttons, title = self._build_choice_widgets()
+
+        super().__init__(
+            title=title, type="custom", content_cls=content, buttons=buttons, **kwargs
+        )
+
+    # ------------------------------------------------------------------
+    # Building widgets for both modes
+    # ------------------------------------------------------------------
+    def _build_select_widgets(self):
+        metrics = core.get_all_metric_types()
+        existing = {m.get("name") for m in self.screen.exercise_obj.metrics}
+        metrics = [
+            m
+            for m in metrics
+            if m["name"] not in existing and m.get("scope") in ("set", "exercise")
+        ]
+        list_view = MDList()
+        for m in metrics:
+            item = OneLineListItem(text=m["name"])
+            item.bind(on_release=lambda inst, name=m["name"]: self.add_metric(name))
+            list_view.add_widget(item)
+
+        scroll = ScrollView(do_scroll_y=True, size_hint_y=None, height=dp(400))
+        scroll.add_widget(list_view)
+
+        new_btn = MDRaisedButton(
+            text="New Metric", on_release=self.show_new_metric_form
+        )
+        cancel_btn = MDRaisedButton(text="Cancel", on_release=lambda *a: self.dismiss())
+        buttons = [new_btn, cancel_btn]
+        return scroll, buttons, "Select Metric"
+
+    def _build_new_metric_widgets(self):
+        default_height = dp(48)
+        self.input_widgets = {}
+
+        schema = core.get_metric_type_schema()
+        if not schema:
+            schema = [
+                {"name": "name"},
+                {"name": "description"},
+                {
+                    "name": "type",
+                    "options": ["int", "float", "str", "bool", "enum", "slider"],
+                },
+                {
+                    "name": "input_timing",
+                    "options": [
+                        "preset",
+                        "pre_session",
+                        "post_session",
+                        "pre_set",
+                        "post_set",
+                    ],
+                },
+                {"name": "scope", "options": ["session", "section", "exercise", "set"]},
+                {"name": "is_required"},
+            ]
+        else:
+            order_map = {field["name"]: field for field in schema}
+            schema = [
+                order_map[name] for name in METRIC_FIELD_ORDER if name in order_map
+            ] + [field for field in schema if field["name"] not in METRIC_FIELD_ORDER]
+
+        form = MDBoxLayout(orientation="vertical", spacing="8dp", size_hint_y=None)
+        form.bind(minimum_height=form.setter("height"))
+
+        def enable_auto_resize(text_field: MDTextField):
+            text_field.bind(
+                text=lambda inst, val: setattr(
+                    inst, "height", max(default_height, inst.minimum_height)
+                )
+            )
+
+        for field in schema:
+            name = field["name"]
+            if name == "enum_values_json":
+                continue
+            options = field.get("options")
+            if name == "is_required":
+                row = MDBoxLayout(
+                    orientation="horizontal", size_hint_y=None, height="40dp"
+                )
+                widget = MDCheckbox(size_hint_y=None, height=default_height)
+                row.add_widget(widget)
+                row.add_widget(MDLabel(text="Required"))
+                form.add_widget(row)
+            elif options:
+                widget = Spinner(
+                    text=options[0],
+                    values=options,
+                    size_hint_y=None,
+                    height=default_height,
+                )
+                form.add_widget(widget)
+            else:
+                widget = MDTextField(
+                    hint_text=name.replace("_", " ").title(),
+                    size_hint_y=None,
+                    height=default_height,
+                    multiline=True,
+                )
+                widget.hint_text_font_size = "12sp"
+                enable_auto_resize(widget)
+                form.add_widget(widget)
+
+            self.input_widgets[name] = widget
+
+        # Text box for enum values. This field only appears when the metric type is ``enum``.
+        self.enum_values_field = MDTextField(
+            hint_text="Enum Values (comma separated)",
+            size_hint_y=None,
+            height=default_height,
+            multiline=True,
+        )
+        self.enum_values_field.hint_text_font_size = "12sp"
+        enable_auto_resize(self.enum_values_field)
+
+        def update_enum_visibility(*args):
+            show = self.input_widgets["type"].text == "enum"
+            has_parent = self.enum_values_field.parent is not None
+            if show and not has_parent:
+                form.add_widget(self.enum_values_field)
+            elif not show and has_parent:
+                form.remove_widget(self.enum_values_field)
+
+        def update_enum_filter(*args):
+            metric_type = self.input_widgets["type"].text
+            if metric_type == "int":
+                allowed = string.digits + ","
+            elif metric_type == "float":
+                allowed = string.digits + ",."
+            else:
+                allowed = string.ascii_letters + " ,"
+
+            def _filter(value, from_undo):
+                filtered = "".join(ch for ch in value if ch in allowed)
+                return re.sub(r",\s+", ",", filtered)
+
+            self.enum_values_field.input_filter = _filter
+
+        if "type" in self.input_widgets:
+            self.input_widgets["type"].bind(
+                text=lambda *a: (update_enum_visibility(), update_enum_filter())
+            )
+            update_enum_visibility()
+            update_enum_filter()
+
+        layout = ScrollView(do_scroll_y=True, size_hint_y=None, height=dp(400))
+        layout.add_widget(form)
+
+        save_btn = MDRaisedButton(text="Save", on_release=self.save_metric)
+        back_btn = MDRaisedButton(text="Back", on_release=lambda *a: self.dismiss())
+        buttons = [save_btn, back_btn]
+        return layout, buttons, "New Metric"
+
+    def _build_choice_widgets(self):
+        label = MDLabel(text="Choose an option", halign="center")
+        add_btn = MDRaisedButton(text="Add Metric", on_release=self.show_metric_list)
+        new_btn = MDRaisedButton(
+            text="New Metric", on_release=self.show_new_metric_form
+        )
+        cancel_btn = MDRaisedButton(text="Cancel", on_release=lambda *a: self.dismiss())
+        content = MDBoxLayout(orientation="vertical", spacing="8dp")
+        content.add_widget(label)
+        buttons = [add_btn, new_btn, cancel_btn]
+        return content, buttons, "Metric Options"
+
+    # ------------------------------------------------------------------
+    # Mode switching helpers
+    # ------------------------------------------------------------------
+    def show_new_metric_form(self, *args):
+        self.dismiss()
+        popup = AddMetricPopup(self.screen, popup_mode="new", mode=self.mode)
+        popup.open()
+
+    def show_metric_list(self, *args):
+        self.dismiss()
+        popup = AddMetricPopup(self.screen, popup_mode="select", mode=self.mode)
+        popup.open()
+
+    def add_metric(self, name, *args):
+        metric_defs = core.get_all_metric_types()
+        for m in metric_defs:
+            if m["name"] == name:
+                self.screen.exercise_obj.add_metric(m)
+                break
+        self.dismiss()
+        self.screen.populate()
+        self.screen.save_enabled = self.screen.exercise_obj.is_modified()
+
+    def save_metric(self, *args):
+        errors = []
+
+        name = self.input_widgets["name"].text.strip()
+        metric_type = self.input_widgets["type"].text
+
+        if not name:
+            errors.append("name")
+
+        existing_names = {m.get("name") for m in self.screen.exercise_obj.metrics}
+        if name and name in existing_names:
+            errors.append("name")
+            if hasattr(self.input_widgets["name"], "helper_text"):
+                self.input_widgets["name"].helper_text = "Duplicate name"
+                self.input_widgets["name"].helper_text_mode = "on_error"
+
+        values = []
+        if metric_type == "enum":
+            text = self.enum_values_field.text.strip()
+            if not text:
+                errors.append("enum_values")
+            else:
+                values = [v.strip() for v in text.split(",") if v.strip()]
+                if not values:
+                    errors.append("enum_values")
+
+        red = (1, 0, 0, 1)
+        for key, widget in self.input_widgets.items():
+            if isinstance(widget, Spinner):
+                widget.text_color = red if key in errors else (1, 1, 1, 1)
+            elif isinstance(widget, MDTextField):
+                widget.error = key in errors
+        self.enum_values_field.error = "enum_values" in errors
+
+        if errors:
+            return
+
+        metric = {}
+        for key, widget in self.input_widgets.items():
+            if isinstance(widget, MDCheckbox):
+                metric[key] = bool(widget.active)
+            else:
+                metric[key] = widget.text
+        metric_type = metric.pop("type", metric_type)
+        metric["type"] = metric_type
+        if values:
+            metric["values"] = values
+
+        db_path = DEFAULT_DB_PATH
+        try:
+            core.add_metric_type(
+                metric["name"],
+                metric["type"],
+                metric["input_timing"],
+                metric["scope"],
+                metric.get("description", ""),
+                metric.get("is_required", False),
+                metric.get("values"),
+                db_path=db_path,
+            )
+        except sqlite3.IntegrityError:
+            self.input_widgets["name"].error = True
+            return
+
+        app = MDApp.get_running_app()
+        if app:
+            app.metric_library_version += 1
+
+        self.screen.exercise_obj.add_metric(metric)
+        self.screen.populate()
+        self.screen.save_enabled = self.screen.exercise_obj.is_modified()
+        self.show_metric_list()
+
+
