+++ ui/popups/constants.py
@@ -1,0 +1,8 @@
+METRIC_FIELD_ORDER = [
+    "name",
+    "description",
+    "type",
+    "input_timing",
+    "scope",
+    "is_required",
+]
